+++ cantera/trunk/include/cantera/thermo/MolalityVPSSTP.h
@@ -810,7 +810,7 @@ public:
      * @param dlnActCoeffdlnN    Output vector of derivatives of the
      *                           log Activity Coefficients. length = m_kk * m_kk
      */
-    virtual void getdlnActCoeffdlnN(const int ld, doublereal* const dlnActCoeffdlnN) {
+    virtual void getdlnActCoeffdlnN(const size_t ld, doublereal* const dlnActCoeffdlnN) {
         getdlnActCoeffdlnN_numderiv(ld, dlnActCoeffdlnN);
     }
